Added built-in Vanilla preset, alphabetized preset list, modicon files
The alphabetized preset list starts with the built-in presets (just the Vanilla preset so far) and the rest follow in alphabetical order.
Our built-in Vanilla preset is just a preset that has zero mods enabled by default. It cannot be changed or deleted.
Also added modicon and preview images for publication.